# backend/openapi.yaml
openapi: 3.0.3
info:
  title: PeajeMexico API
  version: 1.1.0
  description: >
    API backend de PeajeMexico: calculadora de costos de casetas para rutas
    en la red carretera federal de México, respaldada por la API de Ruteo del
    INEGI. Incluye autenticación de usuarios, cálculo de rutas/costos y
    contacto por correo.


    **Estado de montaje (2026-08-01):** no todos los routers definidos en el
    código están activos en `server.js`. Cada operación de abajo indica su
    estado real vía la extensión `x-status`:
    - `live`: montado y accesible hoy.
    - `not-mounted`: el archivo de rutas existe y está completamente
      implementado, pero su `app.use(...)` está comentado en `server.js`, por
      lo que el endpoint no responde en ningún ambiente actualmente.


    **Known issues:**
    - Los prefijos reales de montaje son `/api/ds/...` (ver `server.js`), no
      `/api/users/...` / `/api/mail/...` / `/api/uploads/...` como sugería una
      versión previa de este documento.
    - `user.routes.js` (ahora `live`) y `upload.routes.js` (aún
      `not-mounted`) referencian los roles `store_admin` y `catalog_manager`
      en `authorizeRoles(...)`, pero `middleware/authMiddleware.js` (fuente
      de verdad) solo define los roles `super_admin` y `customer` (igual que
      el enum de `models/user.model.js`). Como `user.routes.js` ya está
      montado, esto es un bug activo hoy, no solo teórico: `GET
...
      estas rutas deben exigir solo `super_admin`.
...
servers:
...
paths:
...
  /api/ds/users/register:
    post:
      tags: [Auth]
      summary: Registrar usuario customer
      x-status: live
      description: >
        Envía correo de verificación con un token `email_verification`
        (24h) hacia `${FRONTEND_URL}/users/verify?token=...`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/EmailAlreadyRegistered"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ds/users/verify:
    get:
      tags: [Auth]
      summary: Verificar cuenta por token de correo
      x-status: live
      parameters:
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cuenta verificada o ya verificada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Token faltante, inválido o expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/UserNotFound"
  /api/ds/users/forgot-password:
    post:
      tags: [Auth]
      summary: Solicitar enlace de recuperación de contraseña
      x-status: live
      description: >
        Envía token `reset_password` hacia
        `${FRONTEND_URL}/reset-password?token=...`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Enlace de recuperación enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ds/users/reset-password:
    post:
      tags: [Auth]
      summary: Cambiar contraseña con token de reset
      x-status: live
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Contraseña cambiada correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Token faltante/inválido/expirado o contraseña inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/UserNotFound"
  /api/ds/users/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      x-status: live
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cuenta no verificada o rol no soportado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ds/users:
    get:
      tags: [Users]
      summary: Listar usuarios
      x-status: live
      x-known-issue: >
        Requiere rol store_admin o super_admin; store_admin no existe en
        authMiddleware.js hoy, así que este endpoint responde siempre 500
        RBAC_CONFIGURATION_INVALID sin importar el rol de quien llame (ver
        Known issues en info.description).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ds/users/{id}:
    get:
      tags: [Users]
      summary: Obtener usuario por id (self o admin)
      x-status: live
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      tags: [Users]
      summary: Actualizar usuario (datos generales y/o rol)
      x-status: live
      x-known-issue: >
        Permite asignar rol store_admin, que no existe en authMiddleware.js
        hoy. Además, si quien llama no es el propio usuario editado, la
        autorización cae en authorizeRoles(store_admin, super_admin) y
        responde siempre 500 RBAC_CONFIGURATION_INVALID (ver Known issues en
        info.description).
      description: >
        No permite cambiar email. Soporta multipart/form-data para
        profileImage. El archivo se valida por contenido real (no solo
        mimetype) y se aceptan únicamente JPG/PNG.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Users]
      summary: Eliminar usuario por id
      x-status: live
      x-known-issue: >
        Requiere rol store_admin o super_admin; store_admin no existe en
        authMiddleware.js hoy, así que este endpoint responde siempre 500
        RBAC_CONFIGURATION_INVALID sin importar el rol de quien llame (ver
        Known issues en info.description).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      responses:
        "200":
          description: Usuario eliminado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteUserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/ds/users/{id}/password:
    patch:
      tags: [Users]
      summary: Cambiar contraseña del usuario autenticado (solo dueño)
      x-status: live
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Contraseña actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/UserNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------
  # Uploads (not-mounted) — /api/ds/uploads
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT de acceso (tokenType=access) firmado con HS256.
        Validación estricta de issuer (JWT_ISSUER) y audience (JWT_AUDIENCE).
  parameters:
    UserIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: ObjectId de usuario
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Sin permisos para la operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UserNotFound:
      description: Usuario no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    EmailAlreadyRegistered:
      description: El correo ya está registrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Demasiadas solicitudes en una ventana de tiempo
      headers:
        Retry-After:
          description: Segundos recomendados para reintentar
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Límite de solicitudes para la ventana
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Solicitudes restantes en la ventana actual
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch en segundos cuando se reinicia la ventana
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            rateLimitExceeded:
              summary: Ejemplo de límite excedido
              value:
                ok: false
                error:
                  status: 429
                  code: RATE_LIMIT_LOGIN_EXCEEDED
                  message: Demasiados intentos de inicio de sesión. Intenta nuevamente más tarde.
                  details:
                    retryAfterSec: 120
...
  schemas:
    Role:
      type: string
      description: >
        Roles soportados hoy por middleware/authMiddleware.js y
        models/user.model.js. No incluye store_admin/catalog_manager
        referenciados por rutas not-mounted (ver Known issues).
      enum:
        - super_admin
        - customer
    UserPublic:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        profileImage:
          type: string
          nullable: true
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [_id, name, email, role, isVerified, createdAt]
    UserListItem:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [_id, name, email, role, isVerified, createdAt]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [name, email, password]
    RegisterResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [message, user]
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [token, newPassword]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [message, token, user]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        role:
          $ref: "#/components/schemas/Role"
        profileImage:
          type: string
          format: binary
          description: Imagen real en formato JPG o PNG (máximo 5MB).
    UpdateUserResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [message, user]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
    DeleteUserResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserPublic"
      required: [message, user]
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            status:
              type: integer
              example: 400
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              oneOf:
                - type: string
                - type: object
                - type: array
                  items: {}
          required: [status, code, message]
      required: [ok, error]